Rename module `citric.response` to `citric.method`
