+++ src/citric/method.py
